requirements.txt
xarray>=2023.1.0
netCDF4>=1.6.0
matplotlib>=3.7.0
cartopy>=0.21.0
numpy>=1.24.0
pyyaml>=6.0.1
tqdm>=4.65.0
imageio>=2.31.0
rasterio>=1.3.0  # For GeoTIFF support